requirements.txt
groq
chromadb
sentence-transformers
pymupdf
Pillow
requests
beautifulsoup4
lxml
python-dotenv